+++ Classes/NotificationCell.m
@@ -13,6 +13,8 @@
 	self = [super initWithStyle:style reuseIdentifier:reuseIdentifier];
 	self.textLabel.font = [UIFont systemFontOfSize:15];
 	self.detailTextLabel.font = [UIFont systemFontOfSize:13];
+	self.textLabel.backgroundColor = [UIColor clearColor];
+	self.detailTextLabel.backgroundColor = [UIColor clearColor];
 	self.selectionStyle = UITableViewCellSelectionStyleBlue;
 	UIImage *accessoryImage = [UIImage imageNamed:@"Remove.png"];
 	UIButton *accessoryButton = [UIButton buttonWithType:UIButtonTypeCustom];
@@ -37,6 +39,23 @@
     UITableView *tableView = (UITableView*)self.superview;
     NSIndexPath *indexPath = [tableView indexPathForCell:self];
     [tableView.delegate tableView:tableView accessoryButtonTappedForRowWithIndexPath:indexPath];
+}
+
+- (void)setCustomBackgroundColor:(UIColor *)color {
+	if (!self.backgroundView) {
+		self.backgroundView = [[UIView alloc] initWithFrame:CGRectZero];
+	}
+	self.backgroundView.backgroundColor = color;
+}
+
+- (void)markAsNew {
+	UIColor *highlightColor = [UIColor colorWithHue:0.6 saturation:0.09 brightness:1.0 alpha:1.0];
+	[self setCustomBackgroundColor:highlightColor];
+}
+
+- (void)markAsRead {
+	UIColor *normalColor = [UIColor whiteColor];
+	[self setCustomBackgroundColor:normalColor];
 }
 
 @end